+++ internal/store/event_store.go
@@ -1,0 +1,9 @@
+package store
+
+type NoopEventStore struct{}
+
+var _ EventStore = (*NoopEventStore)(nil)
+
+func (s *NoopEventStore) Append(Event) error {
+	return nil
+}